Test/nhom 1

+++ HRM/HRM/Rpc/academic-level/AcademicLevelController.cs
@@ -38,6 +38,8 @@ namespace HRM.Rpc.academic_level
         {
             if (!ModelState.IsValid)
                 throw new BindException(ModelState);
+            if (!await HasPermission(AcademicLevel_AcademicLevelDTO.Id))
+                return StatusCode(403);
 
             AcademicLevel AcademicLevel = ConvertDTOToEntity(AcademicLevel_AcademicLevelDTO);
             AcademicLevel = await AcademicLevelService.Create(AcademicLevel);
@@ -54,8 +56,8 @@ namespace HRM.Rpc.academic_level
         {
             if (!ModelState.IsValid)
                 throw new BindException(ModelState);
-            // if (!await HasPermission(AcademicLevel_AcademicLevelDTO.Id))
-            //     return Forbid();
+            if (!await HasPermission(AcademicLevel_AcademicLevelDTO.Id))
+                return StatusCode(403);
 
             AcademicLevel AcademicLevel = ConvertDTOToEntity(AcademicLevel_AcademicLevelDTO);
             AcademicLevel = await AcademicLevelService.Get(AcademicLevel.Id);
@@ -83,6 +85,8 @@ namespace HRM.Rpc.academic_level
         {
             if (!ModelState.IsValid)
                 throw new BindException(ModelState);
+            if (!await HasPermission(AcademicLevel_AcademicLevelDTO.Id))
+               return StatusCode(403);
 
             AcademicLevel AcademicLevel = ConvertDTOToEntity(AcademicLevel_AcademicLevelDTO);
             AcademicLevel = await AcademicLevelService.Update(AcademicLevel);
@@ -95,6 +99,8 @@ namespace HRM.Rpc.academic_level
         {
             if (!ModelState.IsValid)
                 throw new BindException(ModelState);
+            if (!await HasPermission(AcademicLevel_AcademicLevelDTO.Id))
+                return StatusCode(403);
 
             AcademicLevel AcademicLevel = ConvertDTOToEntity(AcademicLevel_AcademicLevelDTO);
             AcademicLevel = await AcademicLevelService.Delete(AcademicLevel);
